Chore: "input 컴포넌트 사용 예시를 inputs 파일에 위치"
